Split acceptance tests and unit tests
