+++ azuredevops/data_agentpools_test.go
@@ -7,19 +7,14 @@ package azuredevops
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/microsoft/azure-devops-go-api/azuredevops/taskagent"
 	"github.com/microsoft/terraform-provider-azuredevops/azdosdkmocks"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/utils/config"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/utils/converter"
-	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/utils/testhelper"
 	"github.com/stretchr/testify/require"
 )
 
@@ -91,70 +86,4 @@ func TestDataSourceAgentPool_Read_TestFindAllAgentPools(t *testing.T) {
 	agentPools := resourceData.Get("agent_pools").([]interface{})
 	require.NotNil(t, agentPools)
 	require.Equal(t, len(dataTestAgentPools), len(agentPools))
-}
-
-/**
- * Begin acceptance tests
- */
-
-// Verifies that the following sequence of events occurrs without error:
-//	(1) TF can create a project
-//	(2) A data source is added to the configuration, and that data source can find the created project
-
-func TestAccAgentPools_DataSource(t *testing.T) {
-	agentPoolName := testhelper.TestAccResourcePrefix + acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	agentPool1Name := agentPoolName + "_1"
-	agentPool2Name := agentPoolName + "_2"
-
-	createAgent1 := testhelper.TestAccAgentPoolResourceAppendPoolNameToResourceName(agentPool1Name)
-	createAgent2 := testhelper.TestAccAgentPoolResourceAppendPoolNameToResourceName(agentPool2Name)
-	agentPoolsData := testhelper.TestAccAgentPoolsDataSource()
-	createAgentPools := fmt.Sprintf("%s\n%s", createAgent1, createAgent2)
-
-	tfNode := "data.azuredevops_agent_pools.pools"
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { testhelper.TestAccPreCheck(t, nil) },
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: createAgentPools,
-			},
-			{
-				Config: agentPoolsData,
-				Check: resource.ComposeTestCheckFunc(
-					testAgentPoolExists(tfNode, agentPool1Name),
-					testAgentPoolExists(tfNode, agentPool2Name),
-				),
-			},
-		},
-	})
-}
-
-func testAgentPoolExists(tfNode string, poolName string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rootModule := s.RootModule()
-		resource, ok := rootModule.Resources[tfNode]
-		if !ok {
-			return fmt.Errorf("Did not find a project in the TF state")
-		}
-
-		is := resource.Primary
-		if is == nil {
-			return fmt.Errorf("No primary instance: %s in %s", tfNode, rootModule.Path)
-		}
-		if !containsValue(is.Attributes, poolName) {
-			return fmt.Errorf("%s does not contain a pool with name %s", tfNode, poolName)
-		}
-		return nil
-	}
-}
-
-func containsValue(m map[string]string, v string) bool {
-	for _, x := range m {
-		if x == v {
-			return true
-		}
-	}
-	return false
 }
